# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Playero API",
        "description": "API oficial do sistema Playero",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/v1/admin/admins": {
            "get": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Listar administradores",
                "operationId": "9e7c3acc4d7d65a6db47933a95d1ad2c",
                "responses": {
                    "200": {
                        "description": "Lista de administradores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Criar administrador",
                "operationId": "a718844c6b4f4abc8ac392e96cfd1f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@playero.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Administrador criado"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/admins/{id}": {
            "put": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Editar administrador",
                "operationId": "ce4d247a97537c696a8dbca5601caa19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@playero.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Administrador atualizado"
                    },
                    "403": {
                        "description": "Sem permissao"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Excluir administrador",
                "operationId": "3e6cca553a5061372d079ed2801eed7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Administrador removido"
                    },
                    "403": {
                        "description": "Sem permissao"
                    },
                    "404": {
                        "description": "Administrador nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/admins/{id}/status": {
            "patch": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Ativar ou desativar administrador",
                "operationId": "87e28fc5ab87c6bea43d03629db52358",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado"
                    },
                    "403": {
                        "description": "Sem permissao"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}